Update urls.py - part of app creation and redirect validation
Created app for userprofile and a basic html page for the leaderboard in the leaderboard app, along with fixing the navbar and where to display it, as well as the styles to use. Ensured that the pages redirect to the index.html page where signing up and logging in can be done when the link contains paths which require user information while no user has been authenticated, i.e. ../accounts/dashboard